Added frontend UI

+++ backend/controllers/goalController.js
@@ -1,11 +1,13 @@
 /* A middleware that wraps async functions to catch errors and pass them to the express error handler. */
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+// to make sure every user can delete their own goals
+const User = require("../models/userModel");
 //@desc   Get goals
 //@route  GET /api/goals
 //@access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
 
   res.status(200).json(goals);
 });
@@ -22,10 +24,13 @@ const setGoal = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(200).json(goal);
 });
+
+
 //@desc   Update goal
 //@route  PUT /api/goals/:id
 //@access Private
@@ -35,6 +40,19 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Goal not found");
   }
+const user = await User.findById(req.user.id)
+  /* Checking if the user exists. */
+  if (!user){
+    res.status(404);
+    throw new Error("User not found");
+  }
+  /* Checking if the user owns the goal. */
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true, //create new goal if it doesn't exist
   });
@@ -50,6 +68,19 @@ const deleteGoal = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Goal not found");
   }
+  const user = await User.findById(req.user.id)
+  /* Checking if the user exists. */
+  if (!user){
+    res.status(404);
+    throw new Error("User not found");
+  }
+  /* Checking if the user owns the goal. */
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   await Goal.findByIdAndDelete(req.params.id);
 
   res.status(200).json({ id: req.params.id });
